Remove the usage of Thread.current
The usage of Thread.current is not nessesary as Rails ensure that each
connection is only accessed by a single thread at a time.

The usage of Thread.current causes issues when you have multiple
master_slave_adapters connecting to different databases as the state
stored within Thread.current is shared across the adapters.

Change-Id: I4decc575489037c9a75057d668fe67e04010bc3f
Reviewed-on: https://gerrit.causes.com/4898
Reviewed-by: Lann Martin <lann@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>