Don't allow database permissions to prevent adapter from working
As we run without elevated privileges, we were seeing this exception in
the staging environment:

  Access denied; you need the SUPER,REPLICATION CLIENT privilege for
  this operation: SHOW MASTER STATUS

As seen here:

  https://causes.airbrakeapp.com/errors/43217126

We could either elevate the running privileges we run with in
production, or simply make this fail gracefully by swallowing permission
errors. We've chosen the latter, as it is equivalent to the
(clock-unware) behavior we already had on the Rails 2 branch.

Related story:
  http://www.pivotaltracker.com/story/show/27717153 (fixed)
  ("Eliminate "Access denied; you need the SUPER,REPLICATION CLIENT
  privilege for this operation: SHOW MASTER STATUS" errors on upgrade
  branch on staging")

Change-Id: I759a0a1b3016d4e17aa92c0f8fb345b0295032c0
Reviewed-on: https://gerrit.causes.com/4991
Reviewed-by: Adam Derewecki <adam@causes.com>
Reviewed-by: Lann Martin <lann@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>