Play nicely with adapters of form "mysql*_master_slave"
This commit moves the massage method into the intialize method of the
master_slave_adapter, where the massage()'d config is actually required.

We need to configure our adapter as 'mysql2_master_slave' so that we can run migrations
and other rake tasks. The expected/usual way of using the master_slave_adapter
is to specify:

  adapter: master_slave_adapter
  connection_adapter: mysql2

We need to specify:

  adapter: mysql2_master_slave_adapter
  connection_adapter: mysql2

One restriction is that we need to give it a name that will match the `/mysql/`
regex used by Rails in the `db:*` Rake tasks namespace (see
`activerecord/lib/active_record/railties/databases.rake`).  Without it, things
like migrations and Rake tasks will not work.

Unfortunately the Rails migration code directly hacks the
ActiveRecord::Base.connection, and so this change won't fix migrations, for
example, it will check the `schema_migrations` table on the slave in order to
decide what migrations to run, then run them on the master.  Ideally, it would
do all of this on the master alone.  In practice, we should be ok, and if we
want to be truly resilient we can make a "migration" environment that doesn't
have master-slave stuff configured at all, forcing migrations to always run on
the master.

Change-Id: I70242c48dc6f770c8ae3911a105a6feff8a56aae
Reviewed-on: https://gerrit.causes.com/16881
Reviewed-by: John Skopis <john.skopis@causes.com>
Tested-by: John Skopis <john.skopis@causes.com>