Add missing :exec_query delegate
The :exec_query method is defined in the AbstractAdapter. The
MasterSlaveAdapter does not implement the method.

Around the time of rails 3.2.13 the :pluck method was refactored to use
exec_query [1]. As such a combination of a rails version > 3.2.13 with
the master slave adapter would not work.

The fix here is incredibly simple, delagate the :exec_query method to
the underlying master connection. Unfortunately, this has the unintented
consequence of sending all pluck()'d queries to the master.

It might make sense to make the :pluck method use :select or similar.

[1] https://github.com/rails/rails/issues/9777

Change-Id: Id8079ae0d443ff1d86d3548f71010217f8138684
Reviewed-on: https://gerrit.causes.com/31670
Reviewed-by: Shane da Silva <shane@causes.com>
Reviewed-by: John Skopis <john.skopis@causes.com>
Tested-by: John Skopis <john.skopis@causes.com>